Rebuilt the test server completely.
Now it has a separate (single) message stream for each client, not 3 global message streams.
Thx to this clients get messages in correct order.
Note: acknowledgements and subscriptionUpdates are ignored completely.
Note: callInfoUpdate and rosterUpdate messages are sent with no regard to which call they belong to - everything's sent to each subscriber.